Implemented URL generation for projects
